Insert the data migration in the transaction
